# Speedup ntpath.lexists