--- a/Misc/NEWS.d/next/Core and Builtins/2024-04-13-11-30-09.gh-issue-117841.eW4w_y.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-04-13-11-30-09.gh-issue-117841.eW4w_y.rst
@@ -1,0 +1,1 @@
+Speedup :func:`os.path.lexists` on Windows.