--- a/Modules/posixmodule.c
+++ b/Modules/posixmodule.c
@@ -5280,6 +5280,86 @@ os__path_isfile_impl(PyObject *module, PyObject *path)
 }
 
 
+static PyObject *
+nt_exists(PyObject *path, int follow_symlinks)
+{
+    path_t _path = PATH_T_INITIALIZE("exists", "path", 0, 1);
+    HANDLE hfile;
+    int result = 0;
+
+    if (!path_converter(path, &_path)) {
+        path_cleanup(&_path);
+        if (PyErr_ExceptionMatches(PyExc_ValueError)) {
+            PyErr_Clear();
+            Py_RETURN_FALSE;
+        }
+        return NULL;
+    }
+
+    Py_BEGIN_ALLOW_THREADS
+    if (_path.fd != -1) {
+        hfile = _Py_get_osfhandle_noraise(_path.fd);
+        if (hfile != INVALID_HANDLE_VALUE) {
+            result = 1;
+        }
+    }
+    else if (_path.wide) {
+        BOOL slow_path = TRUE;
+        FILE_STAT_BASIC_INFORMATION statInfo;
+        if (_Py_GetFileInformationByName(_path.wide, FileStatBasicByNameInfo,
+                                         &statInfo, sizeof(statInfo)))
+        {
+            if (!(statInfo.FileAttributes & FILE_ATTRIBUTE_REPARSE_POINT) ||
+                !follow_symlinks &&
+                IsReparseTagNameSurrogate(statInfo.ReparseTag))
+            {
+                slow_path = FALSE;
+                result = 1;
+            }
+        } else if (_Py_GetFileInformationByName_ErrorIsTrustworthy(
+                        GetLastError()))
+        {
+            slow_path = FALSE;
+        }
+        if (slow_path) {
+            STRUCT_STAT st;
+            if (!follow_symlinks) {
+                if (!LSTAT(_path.wide, &st)) {
+                    result = 1;
+                }
+            }
+            else {
+                hfile = CreateFileW(_path.wide, FILE_READ_ATTRIBUTES, 0, NULL,
+                                    OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS,
+                                    NULL);
+                if (hfile != INVALID_HANDLE_VALUE) {
+                    CloseHandle(hfile);
+                    result = 1;
+                }
+                else {
+                    switch (GetLastError()) {
+                    case ERROR_ACCESS_DENIED:
+                    case ERROR_SHARING_VIOLATION:
+                    case ERROR_CANT_ACCESS_FILE:
+                    case ERROR_INVALID_PARAMETER:
+                        if (!STAT(_path.wide, &st)) {
+                            result = 1;
+                        }
+                    }
+                }
+            }
+        }
+    }
+    Py_END_ALLOW_THREADS
+
+    path_cleanup(&_path);
+    if (result) {
+        Py_RETURN_TRUE;
+    }
+    Py_RETURN_FALSE;
+}
+
+
 /*[clinic input]
 os._path_exists
 
@@ -5293,76 +5373,24 @@ static PyObject *
 os__path_exists_impl(PyObject *module, PyObject *path)
 /*[clinic end generated code: output=f508c3b35e13a249 input=380f77cdfa0f7ae8]*/
 {
-    HANDLE hfile;
-    BOOL close_file = TRUE;
-    path_t _path = PATH_T_INITIALIZE("exists", "path", 0, 1);
-    int result;
-    BOOL slow_path = TRUE;
-    FILE_STAT_BASIC_INFORMATION statInfo;
+    return nt_exists(path, 1);
+}
 
-    if (!path_converter(path, &_path)) {
-        path_cleanup(&_path);
-        if (PyErr_ExceptionMatches(PyExc_ValueError)) {
-            PyErr_Clear();
-            Py_RETURN_FALSE;
-        }
-        return NULL;
-    }
 
-    Py_BEGIN_ALLOW_THREADS
-    if (_path.wide) {
-        if (_Py_GetFileInformationByName(_path.wide, FileStatBasicByNameInfo,
-                                         &statInfo, sizeof(statInfo))) {
-            if (!(statInfo.FileAttributes & FILE_ATTRIBUTE_REPARSE_POINT)) {
-                slow_path = FALSE;
-                result = 1;
-            }
-        } else if (_Py_GetFileInformationByName_ErrorIsTrustworthy(GetLastError())) {
-                    slow_path = FALSE;
-                    result = 0;
-        }
-    }
-    if (slow_path) {
-        if (_path.fd != -1) {
-            hfile = _Py_get_osfhandle_noraise(_path.fd);
-            close_file = FALSE;
-        }
-        else {
-            hfile = CreateFileW(_path.wide, FILE_READ_ATTRIBUTES, 0, NULL,
-                                OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS, NULL);
-        }
-        if (hfile != INVALID_HANDLE_VALUE) {
-            result = 1;
-            if (close_file) {
-                CloseHandle(hfile);
-            }
-        }
-        else {
-            STRUCT_STAT st;
-            switch (GetLastError()) {
-            case ERROR_ACCESS_DENIED:
-            case ERROR_SHARING_VIOLATION:
-            case ERROR_CANT_ACCESS_FILE:
-            case ERROR_INVALID_PARAMETER:
-                if (STAT(_path.wide, &st)) {
-                    result = 0;
-                }
-                else {
-                    result = 1;
-                }
-                break;
-            default:
-                result = 0;
-            }
-        }
-    }
-    Py_END_ALLOW_THREADS
+/*[clinic input]
+os._path_lexists
 
-    path_cleanup(&_path);
-    if (result) {
-        Py_RETURN_TRUE;
-    }
-    Py_RETURN_FALSE;
+    path: 'O'
+
+Test whether a path exists.  Returns True for broken symbolic links
+
+[clinic start generated code]*/
+
+static PyObject *
+os__path_lexists_impl(PyObject *module, PyObject *path)
+/*[clinic end generated code: output=b9a42a50b1df6651 input=5f14942c64196ba2]*/
+{
+    return nt_exists(path, 0);
 }
 
 
@@ -16843,6 +16871,7 @@ static PyMethodDef posix_methods[] = {
     OS__PATH_ISFILE_METHODDEF
     OS__PATH_ISLINK_METHODDEF
     OS__PATH_EXISTS_METHODDEF
+    OS__PATH_LEXISTS_METHODDEF
 
     OS__SUPPORTS_VIRTUAL_TERMINAL_METHODDEF
     {NULL,              NULL}            /* Sentinel */