### Benchmark

```bat
::test.bat
@echo off
echo existent && python -m timeit -s "import ntpath" "ntpath.lexists('test.bat')" && PCbuild\amd64\python.exe -m timeit -s "import ntpath" "ntpath.lexists('test.bat')"
echo non-existent && python -m timeit -s "import ntpath" "ntpath.lexists('foo.bar')" && PCbuild\amd64\python.exe -m timeit -s "import ntpath" "ntpath.lexists('foo.bar')"
```

```none
existent 
5000 loops, best of 5: 51.7 usec per loop # before
5000 loops, best of 5: 48.4 usec per loop # after
# -> 1.07x faster
non-existent
10000 loops, best of 5: 21.5 usec per loop # before
20000 loops, best of 5: 16.5 usec per loop # after
# -> 1.30x faster
```