@nineteendo 2024-04-13
@serhiy-storchaka, is this better? This speeds us `ntpath.lexists()` without exposing a new parameter for `ntpath.exists()`.