@nineteendo 2024-04-14
@barneygale, do you think this can be merged? Or do you want more benchmarks?