@eryksun 2024-04-14
I'm working on what I hope will be an improved version compared to the first draft. AFAIK, the `GetFileInformationByName()` fast path will be generally available when Windows 11 24H2 is released later this year. Until then, and for older systems, we need to focus on improving the 'slow' path. I'm looking to avoid full `STAT` and `LSTAT` calls, accept as a last resort.