@nineteendo 2024-04-15
Are you happy with the new implementation, or do you have more ideas?