@nineteendo 2024-04-15
The performance for existent files has already greatly improved. Nice job!