@eryksun 2024-04-15
> At first glance looks correct, although I am not happy that we added so much complicated code for pure optimization of functions which should not be used in performance critical code (`os.scandir()` should be used instead).

The performance of `os.stat()` and `os.lstat()` will improve on Windows once `GetFileInformationByName()` is generally supported, starting with Windows 11 24H2 later this year. Eventually I expect the builtin `_path_*` tests to be removed when Python stops supporting Windows 10.

A disappointment with these builtin functions is that they can't be leveraged in `pathlib.Path` because they're designed to simply return `True` or `False` without raising `OSError` exceptions. The problem is that the `pathlib.Path` methods `exists()`, `is_dir()`, `is_file()`, and `is_symlink()` only ignore `OSError` exceptions for a small set of `errno` values and Windows error codes.

Regarding `scandir()`, in principle, the C implementations of `_path_isdir()` and `_path_isfile()` could be used by the `DirEntry` methods `is_dir()` and `is_file()`. These methods have to call `STAT()` if the entry is a reparse point, except if `follow_symlinks` is false and it's a name-surrogate reparse point[^1]. That's more work than necessary. However, the `DirEntry` tests only handle `FileNotFoundError` exceptions, so there's a mismatch in error handling. Fortunately there usually aren't so many reparse points that the accumulated cost of `STAT()` calls would matter much.

[^1]: That's if `DirEntry` was updated to behave like `os.stat()` and `os.lstat()`, for which `follow_symlinks` is interpreted liberally to include junctions and other name-surrogate file types, not just literal symlinks, while other types of reparse points are always traversed because they're not meant to be handled like links. For example, `DirEntry.is_dir()` incorrectly returns true for a junction that targets a non-existent path.