@barneygale 2024-04-15
> A disappointment with these builtin functions is that they can't be leveraged in `pathlib.Path` because they're designed to simply return `True` or `False` without raising `OSError` exceptions. The problem is that the `pathlib.Path` methods `exists()`, `is_dir()`, `is_file()`, and `is_symlink()` only ignore `OSError` exceptions for a small set of `errno` values and Windows error codes.

We can at least use it from `Path.glob()`, which suppresses all `OSError`s. I have a draft PR for that here: https://github.com/python/cpython/pull/117858

I wouldn't mind if we made the `is_` methods suppress all `OSError`s rather than a fairly arbitrary subset as we do now. We don't document what's suppressed, and it's changed before.